# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beep communities openapi",
    "description": "API documentation for the Communities service",
    "contact": {
      "name": "communities-core@beep.ovh"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.0.1"
  },
  "paths": {
    "/channels/{channel_id}/messages": {
      "get": {
        "tags": [
          "messages"
        ],
        "operationId": "list_messages",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of messages retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_Message"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal message error"
          }
        }
      }
    },
    "/channels/{channel_id}/messages/search": {
      "get": {
        "tags": [
          "messages"
        ],
        "operationId": "search_messages",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Search query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_Message"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal message error"
          }
        }
      }
    },
    "/messages": {
      "post": {
        "tags": [
          "messages"
        ],
        "operationId": "create_message",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Message created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid message name"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal message error"
          }
        }
      }
    },
    "/messages/attachments": {
      "post": {
        "tags": [
          "attachments"
        ],
        "operationId": "create_attachment",
        "responses": {
          "201": {
            "description": "Attachment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attachment"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid attachment data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal attachment error"
          }
        }
      }
    },
    "/messages/{id}": {
      "get": {
        "tags": [
          "messages"
        ],
        "operationId": "get_message",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Message ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Message is private"
          },
          "404": {
            "description": "Message not found"
          },
          "500": {
            "description": "Internal message error"
          }
        }
      },
      "put": {
        "tags": [
          "messages"
        ],
        "operationId": "update_message",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Message ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Message updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid message name"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not the message owner"
          },
          "404": {
            "description": "Message not found"
          },
          "500": {
            "description": "Internal message error"
          }
        }
      },
      "delete": {
        "tags": [
          "messages"
        ],
        "operationId": "delete_message",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Message ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not the message owner"
          },
          "404": {
            "description": "Message not found"
          },
          "500": {
            "description": "Internal message error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Attachment": {
        "type": "object",
        "required": [
          "id",
          "url"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/AttachmentId"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "AttachmentId": {
        "type": "string",
        "format": "uuid"
      },
      "AuthorId": {
        "type": "string",
        "format": "uuid"
      },
      "ChannelId": {
        "type": "string",
        "format": "uuid"
      },
      "CreateMessageRequest": {
        "type": "object",
        "required": [
          "channel_id",
          "content",
          "attachments"
        ],
        "properties": {
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttachmentId"
            }
          },
          "channel_id": {
            "$ref": "#/components/schemas/ChannelId"
          },
          "content": {
            "type": "string"
          },
          "reply_to_message_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MessageId"
              }
            ]
          }
        }
      },
      "Message": {
        "type": "object",
        "required": [
          "_id",
          "channel_id",
          "author_id",
          "content",
          "attachments",
          "is_pinned",
          "created_at"
        ],
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/MessageId"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttachmentId"
            }
          },
          "author_id": {
            "$ref": "#/components/schemas/AuthorId"
          },
          "channel_id": {
            "$ref": "#/components/schemas/ChannelId"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "is_pinned": {
            "type": "boolean"
          },
          "reply_to_message_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MessageId"
              }
            ]
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "MessageId": {
        "type": "string",
        "format": "uuid"
      },
      "PaginatedResponse_Message": {
        "type": "object",
        "required": [
          "data",
          "total",
          "page"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "_id",
                "channel_id",
                "author_id",
                "content",
                "attachments",
                "is_pinned",
                "created_at"
              ],
              "properties": {
                "_id": {
                  "$ref": "#/components/schemas/MessageId"
                },
                "attachments": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttachmentId"
                  }
                },
                "author_id": {
                  "$ref": "#/components/schemas/AuthorId"
                },
                "channel_id": {
                  "$ref": "#/components/schemas/ChannelId"
                },
                "content": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "is_pinned": {
                  "type": "boolean"
                },
                "reply_to_message_id": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/MessageId"
                    }
                  ]
                },
                "updated_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                }
              }
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "$ref": "#/components/schemas/u64"
          }
        }
      },
      "UpdateMessageRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_pinned": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
      "u64": {
        "type": "integer",
        "format": "int64",
        "minimum": 0
      }
    }
  }
}